# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agency Gateway API
  version: 0.1.0
  description: |
    REST API for the Agency AI agent operating platform. Deploys agents in
    enforced isolation containers with credential scoping, network mediation,
    read-only constraints, and continuous audit.

    All endpoints except /health require API token authentication via the
    X-Agency-Token header. Streaming endpoints accept `Accept: application/x-ndjson`.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8200/api/v1
    description: Local gateway daemon
security:
  - apiToken: []
...
paths:
...
  /ontology/promote:
    post:
      operationId: promoteOntologyCandidate
      summary: Promote an ontology candidate to the base ontology
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
      responses:
        "200":
          description: Candidate promoted
          content:
            application/json:
              schema:
                type: object
                properties:
                  promoted:
                    type: boolean
                  value:
                    type: string
  /ontology/reject:
    post:
      operationId: rejectOntologyCandidate
      summary: Reject an ontology candidate
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
      responses:
        "200":
          description: Candidate rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  rejected:
                    type: boolean
                  value:
                    type: string

  # ── Infrastructure ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: X-Agency-Token
      description: API token generated during `agency init`
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Per-webhook secret for inbound webhook authentication
...